+++ src/styles/MentalHealth.css
@@ -1,7 +1,7 @@
 .mental-health-page {
   min-height: 100vh;
   background: #ffffff;
-  padding: 2rem 0;
+  padding: 4rem 0 4rem 0;
 }
 
 .mental-health-container {
@@ -11,7 +11,8 @@
 }
 
 .mental-health-title {
-  font-size: 5rem;
+  font-family: 'Playfair Display', serif;
+  font-size: 7rem;
   color: #3E4D34;
   text-align: left;
   margin-bottom: 0.8rem;
@@ -37,20 +38,21 @@
 
 .mental-health-card {
   background: #3E4D34;
-  padding: 0.8rem 1rem;
-  border-radius: 30px;
+  padding: 1rem 1.2rem;
+  border-radius: 32px;
   text-decoration: none;
   color: white;
   text-align: center;
   box-shadow: 0 2px 6px rgba(62, 77, 52, 0.2);
   transition: all 0.3s ease;
   border: 2px solid #3E4D34;
-  min-height: 45px;
+  min-height: 55px;
   display: flex;
   align-items: center;
   justify-content: center;
   position: relative;
   overflow: hidden;
+  animation: slideUp 0.4s ease forwards;
 }
 
 .mental-health-card::before {
@@ -76,7 +78,7 @@
 }
 
 .mental-health-card-text {
-  font-size: 0.85rem;
+  font-size: 0.95rem;
   font-weight: 600;
   text-align: center;
   line-height: 1.2;
@@ -85,8 +87,9 @@
 }
 
 /* Specific styling for longer text items */
-.mental-health-card:nth-child(5) .mental-health-card-text {
-  font-size: 0.78rem;
+.mental-health-card:nth-child(5) .mental-health-card-text,
+.mental-health-card:nth-child(7) .mental-health-card-text {
+  font-size: 0.88rem;
 }
 
 /* Responsive Design */
@@ -114,17 +117,18 @@
   }
   
   .mental-health-card {
-    padding: 0.7rem 0.9rem;
-    min-height: 40px;
-    border-radius: 25px;
+    padding: 0.9rem 1.1rem;
+    min-height: 48px;
+    border-radius: 28px;
   }
   
   .mental-health-card-text {
-    font-size: 0.82rem;
+    font-size: 0.87rem;
   }
   
-  .mental-health-card:nth-child(5) .mental-health-card-text {
-    font-size: 0.75rem;
+  .mental-health-card:nth-child(5) .mental-health-card-text,
+  .mental-health-card:nth-child(7) .mental-health-card-text {
+    font-size: 0.8rem;
   }
 }
 
@@ -145,17 +149,18 @@
   }
   
   .mental-health-card {
-    padding: 0.6rem 0.8rem;
-    min-height: 38px;
-    border-radius: 22px;
+    padding: 0.8rem 1rem;
+    min-height: 45px;
+    border-radius: 25px;
   }
   
   .mental-health-card-text {
-    font-size: 0.8rem;
+    font-size: 0.85rem;
   }
   
-  .mental-health-card:nth-child(5) .mental-health-card-text {
-    font-size: 0.72rem;
+  .mental-health-card:nth-child(5) .mental-health-card-text,
+.mental-health-card:nth-child(7) .mental-health-card-text {
+    font-size: 0.78rem;
   }
 }
 
@@ -169,10 +174,6 @@
     opacity: 1;
     transform: translateY(0);
   }
-}
-
-.mental-health-card {
-  animation: slideUp 0.4s ease forwards;
 }
 
 .mental-health-card:nth-child(1) { animation-delay: 0.1s; }
